ASIMD optimization for ARM64 CPU
