+++ internal/bits/min_amd64.go
@@ -1,4 +1,4 @@
-//go:build !purego
+//go:build !purego && amd64
 
 package bits
