+++ internal/bits/min_default.go
@@ -1,4 +1,4 @@
-//go:build !go1.18 && (purego || !amd64)
+//go:build !go1.18 && (purego || !amd64 || !arm64)
 
 package bits
